fix(site): render the terminal chat failure as a transcript row
