+++ site/src/pages/AgentsPage/components/ChatConversation/LiveStreamTail.tsx
@@ -1,3 +1,4 @@
+import { MessageScroller } from "@shadcn/react/message-scroller";
 import { ChatStatusCallout } from "./ChatStatusCallout";
 import type { LiveStatusModel } from "./liveStatusModel";
 
@@ -7,33 +8,35 @@ interface LiveStreamTailContentProps {
 }
 
 // The live assistant turn renders as a timeline row, so the tail below the
-// transcript only carries the empty state and the terminal failure callout.
+// transcript only carries the empty state.
 export const LiveStreamTailContent = ({
 	isTranscriptEmpty,
 	liveStatus,
 }: LiveStreamTailContentProps) => {
-	const terminalStatus = liveStatus.phase === "failed" ? liveStatus : null;
-	const shouldRenderEmptyState =
-		isTranscriptEmpty && liveStatus.phase === "idle";
-
-	if (!shouldRenderEmptyState && !terminalStatus) {
+	if (!isTranscriptEmpty || liveStatus.phase !== "idle") {
 		return null;
 	}
 
 	return (
-		<div
-			className={
-				isTranscriptEmpty
-					? "flex flex-col gap-2"
-					: "mt-2 flex flex-col gap-2 empty:mt-0"
-			}
-		>
-			{shouldRenderEmptyState && (
-				<div className="py-12 text-center text-content-secondary">
-					<p className="text-sm">Start a conversation with your agent.</p>
-				</div>
-			)}
-			{terminalStatus && <ChatStatusCallout status={terminalStatus} />}
+		<div className="py-12 text-center text-content-secondary">
+			<p className="text-sm">Start a conversation with your agent.</p>
 		</div>
+	);
+};
+
+// Terminal failures render as a transcript row so a long error scrolls with
+// the conversation instead of squeezing the composer out of the panel.
+export const TerminalStatusRow = ({
+	liveStatus,
+}: {
+	liveStatus: LiveStatusModel;
+}) => {
+	if (liveStatus.phase !== "failed") {
+		return null;
+	}
+	return (
+		<MessageScroller.Item messageId="terminal-status">
+			<ChatStatusCallout status={liveStatus} />
+		</MessageScroller.Item>
 	);
 };
