+++ site/src/pages/AgentsPage/components/ChatConversation/TerminalStatusRow.stories.tsx
@@ -1,0 +1,353 @@
+import type { Meta, StoryObj } from "@storybook/react-vite";
+import { expect, within } from "storybook/test";
+import { ChatMessageScroller } from "../ChatMessageScroller";
+import { TerminalStatusRow } from "./LiveStreamTail";
+import { buildLiveStatus, pinFixtureClock } from "./storyFixtures";
+
+const meta: Meta<typeof TerminalStatusRow> = {
+	title: "pages/AgentsPage/ChatConversation/TerminalStatusRow",
+	component: TerminalStatusRow,
+	beforeEach: pinFixtureClock,
+	decorators: [
+		(Story) => (
+			<div className="flex h-[320px] flex-col">
+				<ChatMessageScroller
+					hasMoreMessages={false}
+					isFetchingMoreMessages={false}
+					hasFetchMoreError={false}
+					hasFilteredOutRows={false}
+					onFetchMoreMessages={async () => {}}
+				>
+					<Story />
+				</ChatMessageScroller>
+			</div>
+		),
+	],
+};
+export default meta;
+type Story = StoryObj<typeof TerminalStatusRow>;
+
+export const UsageLimitExceeded: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			persistedError: {
+				kind: "usage_limit",
+				message: "Your AI spend budget has been reached.",
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /usage limit reached/i }),
+		).toBeVisible();
+		expect(canvas.getByText(/ai spend budget has been reached/i)).toBeVisible();
+		expect(
+			canvas.queryByRole("link", { name: /view usage/i }),
+		).not.toBeInTheDocument();
+	},
+};
+
+export const ProviderQuotaExceeded: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			streamError: {
+				kind: "usage_limit",
+				message:
+					"The usage quota for OpenAI has been exceeded. Check the billing and quota settings for the provider account.",
+				provider: "openai",
+				retryable: false,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByText(/usage quota for openai has been exceeded/i),
+		).toBeVisible();
+		expect(
+			canvas.queryByRole("link", { name: /view usage/i }),
+		).not.toBeInTheDocument();
+		expect(
+			canvas.getByRole("heading", { name: /usage limit reached/i }),
+		).toBeVisible();
+	},
+};
+
+/** Provider failures keep the footer-level terminal callout and status link. */
+
+export const TerminalOverloadedError: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			persistedError: {
+				kind: "overloaded",
+				message: "Anthropic is temporarily overloaded.",
+				provider: "anthropic",
+				retryable: true,
+				statusCode: 529,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /service overloaded/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(/anthropic is temporarily overloaded\./i),
+		).toBeVisible();
+		expect(canvas.getByText(/^HTTP 529$/)).toBeVisible();
+		expect(canvas.queryByText(/please try again/i)).not.toBeInTheDocument();
+		expect(canvas.queryByText(/^retryable$/i)).not.toBeInTheDocument();
+		expect(canvas.getByRole("link", { name: /status/i })).toBeVisible();
+		expect(canvas.queryByText(/provider anthropic/i)).not.toBeInTheDocument();
+	},
+};
+
+/** Content-filter refusals render as terminal errors without a retry countdown or status link. */
+
+export const TerminalContentFilterError: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			persistedError: {
+				kind: "content_filter",
+				message:
+					"Anthropic blocked this response under its content policy (cyber).",
+				detail:
+					"This request triggered restrictions on violative cyber content and was blocked under Anthropic's Usage Policy. To learn more, see https://platform.claude.com/docs/en/build-with-claude/refusals-and-fallback.",
+				provider: "anthropic",
+				retryable: false,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /response blocked/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(
+				/anthropic blocked this response under its content policy \(cyber\)\./i,
+			),
+		).toBeVisible();
+		expect(
+			canvas.getByText(/this request triggered restrictions/i),
+		).toBeVisible();
+		expect(canvas.queryByText(/retrying in/i)).not.toBeInTheDocument();
+		expect(
+			canvas.queryByRole("link", { name: /status/i }),
+		).not.toBeInTheDocument();
+	},
+};
+
+/**
+ * Transport timeouts render the per-provider "temporarily
+ * unavailable" copy with a "Request timed out" heading rather than
+ * the generic "Request failed" fallback.
+ */
+
+export const TerminalTimeoutErrorAnthropic: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			streamError: {
+				kind: "timeout",
+				message: "Anthropic is temporarily unavailable.",
+				provider: "anthropic",
+				retryable: false,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /request timed out/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(/anthropic is temporarily unavailable/i),
+		).toBeVisible();
+		// Guard against the pre-fix generic fallback.
+		expect(
+			canvas.queryByText(/the chat request failed unexpectedly/i),
+		).not.toBeInTheDocument();
+	},
+};
+
+/** Transport timeout with an unknown provider uses the generic subject. */
+
+export const TerminalTimeoutErrorUnknownProvider: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			streamError: {
+				kind: "timeout",
+				message: "The AI provider is temporarily unavailable.",
+				retryable: false,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /request timed out/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(/the ai provider is temporarily unavailable/i),
+		).toBeVisible();
+	},
+};
+
+/** Missing API key shows the "Chat interrupted" terminal error. */
+
+export const TerminalMissingKeyError: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			streamError: {
+				kind: "missing_key",
+				message:
+					"This conversation was started with an API key that is no longer available. Send your message again to continue.",
+				retryable: false,
+				detail:
+					"If this error persists after resending, please report it as a bug.",
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /chat interrupted/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(
+				/this conversation was started with an api key that is no longer available/i,
+			),
+		).toBeVisible();
+		expect(
+			canvas.getByText(/if this error persists after resending/i),
+		).toBeVisible();
+		// Guard against the generic fallback.
+		expect(
+			canvas.queryByText(/the chat request failed unexpectedly/i),
+		).not.toBeInTheDocument();
+	},
+};
+
+/** Terminal stream-silence timeouts get a specific heading without provider metadata. */
+
+export const TerminalStreamSilenceTimeoutError: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			persistedError: {
+				kind: "stream_silence_timeout",
+				message: "Anthropic did not send response data in time.",
+				provider: "anthropic",
+				retryable: true,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /response stalled/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(/anthropic did not send response data in time./i),
+		).toBeVisible();
+		expect(canvas.queryByText(/please try again/i)).not.toBeInTheDocument();
+		expect(canvas.queryByText(/^retryable$/i)).not.toBeInTheDocument();
+		expect(
+			canvas.queryByRole("link", { name: /status/i }),
+		).not.toBeInTheDocument();
+		expect(canvas.queryByText(/provider anthropic/i)).not.toBeInTheDocument();
+	},
+};
+
+/** Disabled provider errors render an admin-oriented message without retry. */
+
+export const TerminalProviderDisabledError: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			streamError: {
+				kind: "provider_disabled",
+				message:
+					"The OpenAI provider has been disabled. Contact your Coder administrator.",
+				provider: "openai",
+				retryable: false,
+				statusCode: 503,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /provider disabled/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(
+				/the openai provider has been disabled.*contact your coder administrator/i,
+			),
+		).toBeVisible();
+		expect(canvas.getByText(/^HTTP 503$/)).toBeVisible();
+		// No retry or status link for administrative disablement.
+		expect(canvas.queryByText(/retrying/i)).not.toBeInTheDocument();
+		expect(
+			canvas.queryByRole("link", { name: /status/i }),
+		).not.toBeInTheDocument();
+	},
+};
+
+/** Generic failures do not show usage or provider CTAs. */
+
+export const GenericErrorDoesNotShowUsageAction: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			persistedError: {
+				kind: "generic",
+				message: "Provider request failed.",
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /request failed/i }),
+		).toBeVisible();
+		expect(canvas.getByText(/provider request failed/i)).toBeVisible();
+		expect(
+			canvas.queryByText(/start a conversation with your agent/i),
+		).not.toBeInTheDocument();
+		expect(
+			canvas.queryByRole("link", { name: /view usage/i }),
+		).not.toBeInTheDocument();
+		expect(
+			canvas.queryByRole("link", { name: /status/i }),
+		).not.toBeInTheDocument();
+	},
+};
+
+/** Provider detail renders in a monospace block for generic errors. */
+
+export const GenericErrorShowsProviderDetail: Story = {
+	args: {
+		liveStatus: buildLiveStatus({
+			streamError: {
+				kind: "generic",
+				message: "Anthropic returned an unexpected error.",
+				detail:
+					"messages.0.content.1.image.source.base64: image exceeds 5 MB maximum.",
+				provider: "anthropic",
+				statusCode: 400,
+				retryable: false,
+			},
+		}),
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		expect(
+			canvas.getByRole("heading", { name: /request failed/i }),
+		).toBeVisible();
+		expect(
+			canvas.getByText(/anthropic returned an unexpected error\./i),
+		).toBeVisible();
+		expect(canvas.getByText(/^HTTP 400$/)).toBeVisible();
+		expect(canvas.getByText(/image exceeds 5 mb maximum/i)).toBeVisible();
+	},
+};
